장바구니 부분 컨트롤러 구현중

+++ src/main/java/com/jbnu/cocofarm/domain/user/Basket.java
@@ -30,7 +30,7 @@ public class Basket extends BaseTime {
     private Long id;
 
     @Column
-    private Long count;
+    private Integer count;
 
     @ManyToOne(fetch = FetchType.LAZY)
     private User user;
